Create code-3
this code is related to the control of motor using PWM signal it is basically creating a loop of all three components (i.e, motor driver, motor and ardunio). where we are providing a 25 value of PWM signal to all the three commands i.e, forward, backward and rest. so for 200 miliseconds the motor turns forward(dir=1), then backwards to its og postion for next 200ms (dir=-1) and then const (dir=0) no movement.